ref: add another layer of component to handle logic while parent app component will fetch and pass the fetched notes to the main contents
