+++ src/components/LoadingScreen.tsx
@@ -1,0 +1,8 @@
+const LoadingScreen = () => {
+	return (
+		<div id="loading-screen" className="">
+			<div className="">LOADING</div>
+		</div>
+	);
+};
+export default LoadingScreen;
